move docs to lib/api
