+++ .gitignore
@@ -1,4 +1,2 @@
 /build
 /result
-/docs/docs.go
-/docs/swagger.*
