+++ lib/api/docs/.gitignore
@@ -1,0 +1,2 @@
+/docs.go
+/swagger.*
